simple-messenger/requirements.txt
flask>=3.0
werkzeug>=3.0
minio>=7.2
requests>=2.31
psycopg2-binary>=2.9
redis>=5.0
